+++ lib/rubocop/rspec/language/node_pattern.rb
@@ -23,6 +23,9 @@ module RuboCop
 
         def_node_matcher :let?, Helpers::ALL.block_or_block_pass_pattern
 
+        def_node_matcher :include?,
+                         Includes::ALL.send_or_block_or_block_pass_pattern
+
         def_node_matcher :subject?, Subject::ALL.block_pattern
       end
     end